Improved memory file generation scripts.
- Re-arranged the lagic within the genEEPROM.sensorMemLocationToBin function
to simplify search for correct value for the given memory location.
- Renamed some CANID names in CAN.json to be more consistent throughout
config files and easier to parse by generation scripts.
- Modified usage in some sensors in EEPROM.json to remain consistent with
other config files. Added Filter to sensors in EEPROM.json to specify
the biquad filter coefficients to be added to the EEPROM memory files.
- Added FILTERS.json to hold multiple different filters that can be used
for the sensor biquad filters or possibly other types of filters in
the future.