Added usage details to hardware section of ARD
genHARDWARE.py now adds a new entry to the table for a particular
sensor if it is being used for something according to EEPROM.json.
Sensors that are used multiple times will have multiple entries in
the table. A new unused entry has been added to HARDWARE.json for
testing of unused hardware entries.