Added CAN Bus load calculations.
The ARD scripts now calculate the best case and worst case CAN bus
load for the CAN messages in CAN.json. The variability is due to
the possibility of bit stuffing when there are 5 or more of the same
bits in a row.
A new section is created under the 'CAN IDs' section of the ARD
containing the bus loads with a few preset CANBus frequency settings.