Implemented generation of the CANIDs.h header file.
Using the CAN.json file the CANIDs.h header file is created
providing an enum with all of the CANIDs for the OBC firmware
and the SEB firmware. The contents of this file is based on
some test firmware I wrote for the OBC, it is my best guess
as to what will be needed in the final program.