Created OBC firmware
Created firmware that will run on the pocketbeagle. Three threads were
created: the event thread for event handling, eventTimer thread for
adding timers for event handling, and the test thread which is only
used for testing software without dealing with canbus. This has only
been tested on x86 so far however it should work just fine on arm.