Added states section to ARD
Documentation on how each state of the rocket parses each CANID
in Helix-OBC-Firmware is created by Doxygen. The ARD scripts
now parse through this documentation to be included in the ARD.

Some formatting changes were made in the Helix-OBC-Firmware
source files to make parsing easier. generation of the CANIDs.h
header file now uses the STATES.json config file as one of it's
sources.