Implemented execution of state enter and exit functions.
The event parsing loop now calls the enter and exit functions
when entering and leaving states. In order to make management
of the enter and exit functions easier extra "CANIDs" are
added. This should never be called via a received CANID but
rather called explicitly when a transition between states is
required. This choice was made so the initialization of each
state can use the same canParseFunctions array rather than
requiring multiple different arguments to be passed to each
function.