Moved location of script generated header files.
Header files made by the ARD script are now located in the
include directory for each firmware project in this repository.
This change is being made to make software development easier
so that build tools don't have to look in higher level
directories for header files. The same generated header file
should be exactly the same in each project.