Added option to ARD scripts to generate only header files.
We need the ability to generate only header files in order
build the Helix-OBC-Firmware with the most recent changes
to those header files. In addition documentation in those
files should be updated before running the rest of the ard
script. In the current implementation of the makefiles the
headers are generated twice, one fore the firmware and once
again before generating the latex files. While not efficient
this is ok for now.