Implemented make clean in ARD/ and fixed bug in ARD.tex
The clean target in ARD/Makefile now deletes the files
generated by the ARD scripts and xelatex except for the
header files created for the firmware projects.

Since the tex/ and memory/ directories are deleted in
the clean the ARD scripts now need to create those
directories if they don't already exist.

Also fixed a bug in ARD.tex where the wrong tex file was
being included for the EEPROM Layouts section of the ARD.
I believe the tex file that was being included is an old
file that I forgot to delete at some point in time. The
proper file was getting generated by the scripts but
not included in the final ARD.pdf.