Reimplemented eventTimer with a boost circular buffer
In the future I would like at add the ability to remove
particular messages from the eventTimer if for example
there is a state transition and a particular message
should no longer be sent. In order to implement this
feature I will need to iterate through each message
in the queue, deleting it if needed. We cannot iterate
through a c++ priority queue so the eventTimer needed
to move to a different data structure.

Moving to a boost circular buffer has the added benefit
making the eventTimer use only allocate memory statically
as opposed to the priority queue which was based on a
vector and would dynamically allocate memory.