Implemented function to erase messages from the event timer
There is a new function named eventTimerErase which removes
events from the event timer depending on their canid. This
is done by passing a function to the erase function, that
function will be called with the canid for each message in
the event timer queue and if that function returns true the
corresponding message will be deleted. The compare function
can either be a function pointer or a lambda expression.