Implemented example test case where leak check fails
A new #elif section was added to give a test case where the
system fails leak check but should not go to ground safe state.
There is new logic on how the function checkHeliumPressure gets
the most recent pressure reading since that argument to that
function doesn't contain any actual data. The system also does
not return to STATE_IDLE_STATE since it's not implemented yet
and causes a segfault when trying to call the nonexistent
enterIdleState function.