ARD scripts can now be run from the project top level directory
The Bug:
The ARD scripts assume that they are being run from the ARD folder
which would cause the script to break if run from the project root
directory. The make scripts assume that it can be run from the top
level directory.

The Fix:
The ARD script first checks to see if it is begin run from the ARD
folder, if it is it continues on. Otherwise it tries to cd into
the ARD directory. If the script is called from anywhere other
than the project root directory or the ARD folder the script will
fail.