Modified genEEPROM go make code less spaghetti like
Previously when looking for the CANID that corresponds to a given
sensor in EEPROM.json the script would search through CAN.json for
the ID whose CANID_NAME matched the Usage listed in EEPROM.json.
This method was really unreliable since the search only required
the CANID_NAME contain the usage as a substring so mismatches
could easily happen and the two values would have to be changed
at the same time (which would seem unclear to someone modifying
one of the files).

The fix for this is to specify the CANID in the entry in
EEPROM.json such that there would be need need to match the usage
and name. Now only the IDs would need to match which would be
much more obvious. This also has the benefit of making the code
in genEEPROM much simpler.

The EEPROMLAYOUT.json and FILTERS.json files were also changed
to be a little easier to parse in genEEPROM.py.